[Patchset 2] Add TriggerConfig to the TraceConfig proto.
This allows future CLs to use the trigger_config as checked in. Update all the
hashes and other dependencies. The only need stuff is the TriggerConfig and
submessages.

+++ include/perfetto/tracing/core/trace_config.h
@@ -303,17 +303,17 @@ class PERFETTO_EXPORT TraceConfig {
         producer_name_ = value;
       }
 
-      int64_t finalize_trace_delay_ms() const {
+      uint32_t finalize_trace_delay_ms() const {
         return finalize_trace_delay_ms_;
       }
-      void set_finalize_trace_delay_ms(int64_t value) {
+      void set_finalize_trace_delay_ms(uint32_t value) {
         finalize_trace_delay_ms_ = value;
       }
 
      private:
       std::string name_ = {};
       std::string producer_name_ = {};
-      int64_t finalize_trace_delay_ms_ = {};
+      uint32_t finalize_trace_delay_ms_ = {};
 
       // Allows to preserve unknown protobuf fields for compatibility
       // with future versions of .proto files.
